pyproject.toml
[project]
name = "slim-demes"
version = "0.1.0"
description = "Helper Python scripts for slim-demes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "defopt>=6.4.0",
    "demes>=0.2.3",
    "jupyterlab>=4.2.5",
    "matplotlib>=3.9.2",
    "moments-popgen>=1.2.2",
    "msprime>=1.3.3",
    "polars>=1.12.0",
    "pyslim>=1.0.4",
    "pyyaml>=6.0.2",
    "snakemake>=8.24.1",
    "stdpopsim>=0.2.0",
    "timestamp>=0.0.1",
]

[project.scripts]
slimdemes = "slimdemes:main"

